requirements.txt
openai
google-genai

